actually better to split this so opendbc will test docs conventions, and openpilot will test diffing
